refactor: move mirage folder to api
