+++ src/api/mirage/controllers/AuthController.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from "miragejs";
-import { Schema } from "@/mirage/types";
+import { Schema } from "@/api/mirage/types";
 
 class AuthController {
     public login(schema: Schema, request: Request) {
